perf(dsp): major NEON optimization and SWIG Python enhancements
Added:
- SWIG bindings for std::vector/std::array with Python accessors
- Complete Python test suite with 10+ test sections
- sqrtf_neon edge case handling for negative/zero inputs
- UI LED indicators for sweep/burst mode status

Changed:
- Modernized NEON asm: VLD1/VST1 instead of VLDM/VSTM
- Added PLD prefetch hints and Newton-Raphson division refinement
- Updated CMake: -O3, -ftree-vectorize
- Moved NEON math from spectrumpro to api-dsp

Fixed:
- Phase index overflow (int64_t cast)
- Uninitialized fmode variable
- UART JSON source (TX instead of RX)

Removed:
- Obsolete spectrumpro NEON math code
- Dead commented SWIG code"

+++ apps-tools/scopegenpro/src/sig_gen.cpp
@@ -15,9 +15,8 @@ void synthesis_arb(CFloatBinarySignal* signal, const float* data, const GenChann
     double phaseStep = (double)settings.frequency * settings.tscale * 0.01 * settings.arb_size / sigSize;
 
     double currentPhase = 0.0;
-
     for (int i = 0; i < sigSize; ++i) {
-        int z = (int)currentPhase % settings.arb_size;
+        int64_t z = (int64_t)currentPhase % (int64_t)settings.arb_size;
 
         (*signal)[i] = data[z] * settings.amplitude + settings.offset + settings.showOff;
